% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{combineSummaryText}
\alias{combineSummaryText}
\title{Combine multi-response}
\usage{
combineSummaryText(
  screening.text,
  screening.method = c("cor.test", "pearson", "spearman", "kendall", "t.test",
    "t.test.equal", "t.test.unequal", "t.test.paired", "w.test", "w.test.ranksum",
    "w.test.paired", "anova", "anova.equal", "anova.unequal", "kruskal.wallis",
    "chisq.test", "coxph", "moderated.t", "moderated.F", "sam.test", "missing.value",
    "above.median", "above.minimum", "median", "variability"),
  selection.text,
  selection.method = c("cutoff", "rank", "percentile", "fpr", "fdr")
)
}
\arguments{
\item{screening.text}{character string containing the summary of the screening.}

\item{screening.method}{character string indicating the screening method id.}

\item{selection.text}{character string containing the summary of the selection.}

\item{selection.method}{character string indicating the selection method id.}
}
\value{
A character string.
}
\description{
If screened features are from multi-response data,
this function allows to adopt different strategies.
}
\author{
Alessandro Barberis
}
\keyword{internal}
