% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-test-functions.R
\name{coxPH}
\alias{coxPH}
\title{Fit Univariate Proportional Hazards Regression Model}
\usage{
coxPH(x, y, ...)
}
\arguments{
\item{x}{independent variable}

\item{y}{\code{data.frame} with two columns, \code{time} and \code{status}}

\item{...}{further arguments to \code{\link[survival]{coxph}}}
}
\description{
Fits a Cox proportional hazards regression model.
}
\details{
It is a wrapper to \code{\link[survival]{coxph}}
function.
}
\examples{
\dontrun{
#Seed
set.seed(1010)

#Data
x = c(0,2,1,1,1,0,0)
y = data.frame(
  time = c(4,3,1,1,2,2,3),
  status = c(1,1,1,0,1,1,0)
)

#Compute
coxPH(x = x, y = y)
}
}
\author{
Alessandro Barberis
}
\keyword{internal}
