% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-biostats-test-functions.R
\name{createContrastMatrix}
\alias{createContrastMatrix}
\title{Create a matrix of custom contrasts}
\usage{
createContrastMatrix(x, target = NULL, ref = NULL)
}
\arguments{
\item{x}{a factor, a character vector or an object with column names providing
the names of the parameters of which contrasts are desired}

\item{target}{character string indicating the target class}

\item{ref}{character string indicating the reference class}
}
\value{
A numeric matrix representing contrasts between a set of parameters.
}
\description{
Simple function to generate a default contrast matrix from a
design matrix or factor.
See the \strong{Details} section below for further information.
}
\details{
This function is internally used to create a contrast matrix when not
directly provided in the specific cases of binomial and multinomial response.

There are four main settings.

If \code{target} and \code{ref} are provided, the contrast is created so that
the target class is compared against the reference (i.e., target - ref).

If \code{target} is provided and \code{ref = NULL}, the contrast is created so
that the target class is compared against the other classes (i.e., target -
(class1 + class2 + ...)).

If \code{ref} is provided and \code{target = NULL}, the contrasts are created so
that the each class is compared against the reference classes (e.g., class1 -
ref).

If \code{target} and \code{ref} are \code{NULL}, the contrasts are created so
that each class is compared against the remaining classes (e.g., class1 -
(class2 + class3), class2 - (class1 + class3), class3 - (class1 + class2)).
}
\examples{
\dontrun{
#Character vector (binomial)
x = c("a", "b")
createContrastMatrix(x)
createContrastMatrix(x, "a")
createContrastMatrix(x, NULL, "b")
createContrastMatrix(x, "a", "b")

#Character vector
x = c("a", "b", "c")
createContrastMatrix(x)
createContrastMatrix(x, "a")
createContrastMatrix(x, NULL, "b")
createContrastMatrix(x, "a", "b")

#Factor
x = as.factor(sample(c("a", "b", "c"),10,replace=T))
createContrastMatrix(x)
createContrastMatrix(x, "a")
createContrastMatrix(x, NULL, "b")
createContrastMatrix(x, "a", "b")
}
}
\author{
Alessandro Barberis
}
\keyword{internal}
