% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{defaultRanking}
\alias{defaultRanking}
\title{Default Ranking}
\usage{
defaultRanking(
  x,
  id = c("cor.test", "pearson", "spearman", "kendall", "t.test", "t.test.equal",
    "t.test.unequal", "t.test.paired", "w.test", "w.test.ranksum", "w.test.paired",
    "anova", "anova.equal", "anova.unequal", "kruskal.wallis", "chisq.test", "coxph",
    "moderated.t", "moderated.F", "sam.test", "missing.value", "above.median",
    "above.minimum", "median", "variability"),
  multi
)
}
\arguments{
\item{x}{output of the statistical function used for screening.}

\item{id}{character string indicating statistical method used for screening.}

\item{multi}{character string indicating what to do in case of multi-response
(available options are \code{"max"}, \code{"min"}, \code{"avg"}, \code{"sum"})
or integer indicating the column of \code{x} to keep.}
}
\value{
The ordering index vector.
}
\description{
Ranks output of statistical method used for screening.
}
\author{
Alessandro Barberis
}
