% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-stats-functions.R
\name{efficiency}
\alias{efficiency}
\title{Efficiency}
\usage{
efficiency(x, g = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{numerical vector.}

\item{g}{(optional) vector or factor object giving the group for the corresponding
elements of \code{x}.}

\item{na.rm}{logical indicating whether missing values should be removed before
computation.}
}
\value{
A numerical value or a vector containing the computed measure per class.
}
\description{
This function computes the efficiency as the square
of the coefficient of variation.

See the \strong{Details} section below for further information.
}
\details{
The \emph{efficiency} is a measure of dispersion. It is here defined as
the square ratio of the standard deviation to the mean:

\deqn{ EFFICIENCY = (\frac{standard deviation}{mean})^2 = \frac{\sigma^2}{\mu^2}}

If \code{x} can be partitioned into \eqn{c} subgroups (provided by \code{g}),
then the \eqn{EFFICIENCY} is computed for each class.
}
\examples{
#Seed
set.seed(1010)

#Define size
n = 10

#Data
x = sample.int(n = 100, size = n, replace = TRUE)

#Grouping variable
g = c(rep("a", n/2), rep("b", n/2))

#Efficiency
efficiency(x)

#Efficiency by group
efficiency(x = x, g = g)

}
\references{
https://en.wikipedia.org/wiki/Efficiency_(statistics)#Estimators_of_u.i.d._Variables
}
\author{
Alessandro Barberis
}
