% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-featscreen-class.R
\name{featscreen}
\alias{featscreen}
\title{Constructor of the \code{featscreen} class}
\usage{
featscreen(
  method = c("cor.test", "t.test", "w.test", "anova", "kruskal.wallis", "chisq.test",
    "coxph", "moderated.t", "moderated.F", "sam.test", "missing.value", "above.median",
    "above.minimum", "median", "variability"),
  multi = character(),
  selection = c("cutoff", "rank", "percentile", "fpr", "fdr"),
  summary = character(),
  n = integer(),
  features = character(),
  keep = vector(),
  ranks = integer()
)
}
\arguments{
\item{method}{character string, the id of the used screening method.}

\item{multi}{(optional) character string, the id of the used multi-response aggregation method.}

\item{selection}{character string, the id of the used selection method.}

\item{summary}{character string, a textual summary of the screening.}

\item{n}{integer, the dimension of the feature space.}

\item{features}{string vector, the names of the features constituting the feature space.}

\item{keep}{integer or logical vector, the features to keep.}

\item{ranks}{(optional) integer vector, the ordering indices.}
}
\value{
An object of class \code{featscreen}.
}
\description{
This is the constructor of the \code{featscreen} class.
}
\details{
An object of class \code{featscreen} is a named list with eight elements:
\describe{
\item{\code{method}}{the id of the used screening method}
\item{\code{multi}}{the id of the used multi-response aggregation method}
\item{\code{selection}}{the id of the used selection method}
\item{\code{summary}}{a textual summary of the screening}
\item{\code{n}}{the dimension of the feature space}
\item{\code{features}}{the feature names}
\item{\code{keep}}{the features to keep}
\item{\code{ranks}}{the feature ranks}
}

Functions to facilitate access to the data stored in a \code{resampling} object are
available:
\itemize{
\item \code{?getScreeningMethodId}: returns the screening method id
\item \code{?getMultiresponseAggregationMethodId}: returns the multi-response aggregation method id
\item \code{?getSelectionMethodId}: returns the selection method id
\item \code{?getSummary}: returns the textual summary of the screening
\item \code{?getFeatureDimensionality}: returns the dimension of the feature space
\item \code{?getFeatureNames}: returns the feature names
\item \code{?getScreenedFeatures}: returns the features to keep
\item \code{?getFeatureRanks}: returns the feature ranks
}

Other useful functions include:
\itemize{
\item \code{\link[=print.featscreen]{print}}: print a summary of the \code{featscreen} object
}
}
\examples{
#default
featscreen()

#featscreen object
featscreen(
 method = 'cor.test',
 selection = 'cutoff',
 summary = '3 out of 5 features selected by a cutoff.',
 n = 5,
 keep = c(TRUE,FALSE,TRUE,FALSE,TRUE),
 features = paste0("f", seq_len(5)),
 ranks = c(1,5,3,4,2)
)

}
\author{
Alessandro Barberis
}
