% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-biostats-test-functions.R
\name{genericRowBioStat}
\alias{genericRowBioStat}
\title{Stat Function Doc}
\arguments{
\item{x}{\code{matrix} or \code{data.frame}.}

\item{g}{a vector or factor object giving the group for the corresponding
elements of \code{x}.}

\item{y}{numeric vector of data values. Must have the same length as \code{ncol(x)}.}
}
\value{
A list containing two elements:

\describe{
\item{statistic}{A numeric vector, the values of the test statistic}
\item{significance}{A numeric vector, the p-values of the selected test}
}
}
\description{
Stat Function Doc
}
\author{
Alessandro Barberis
}
\keyword{internal}
