% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-test-functions.R
\name{genericRowStat}
\alias{genericRowStat}
\title{Stat Function Doc}
\arguments{
\item{x}{\code{matrix} or \code{data.frame}.}

\item{g}{a vector or factor object giving the group for the corresponding
elements of \code{x}.}

\item{y}{numeric vector of data values. Must have the same length as \code{ncol(x)}.}

\item{alternative}{character string or vector of length \code{nrow(x)}.
The alternative hypothesis for each row of \code{x}.
Values must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"less"}.}

\item{conf.level}{numerical value or numeric vector of length \code{nrow(x)}.
The confidence levels of the intervals.
All values must be in the range of \eqn{[0:1]} or \code{NA}.}

\item{null}{numerical value or numeric vector of length \code{nrow(x)}.
The true values of the difference in means between the two groups of
observations for each row.}
}
\value{
A list containing two elements:

\describe{
\item{statistic}{A numeric vector, the values of the test statistic}
\item{significance}{A numeric vector, the p-values of the selected test}
}
}
\description{
This function computes a statistical measure from each row of
the input data and the output variable.
}
\author{
Alessandro Barberis
}
\keyword{internal}
