% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{getDefaultRankingOptions}
\alias{getDefaultRankingOptions}
\title{Default Ranking Options}
\usage{
getDefaultRankingOptions(
  id = c("cor.test", "pearson", "spearman", "kendall", "t.test", "t.test.equal",
    "t.test.unequal", "t.test.paired", "w.test", "w.test.ranksum", "w.test.paired",
    "anova", "anova.equal", "anova.unequal", "kruskal.wallis", "chisq.test", "coxph",
    "moderated.t", "moderated.F", "sam.test", "missing.value", "above.median",
    "above.minimum", "median", "variability")
)
}
\arguments{
\item{id}{character string, one of the supported statistical techniques.}
}
\value{
A list with three elements:
\describe{
\item{measure}{A character string indicating whether to use the statistic or the significance (where computed)}
\item{order}{A character string indicating whether to sort the statistic/significance in ascending or descending order}
\item{use.abs}{A logical value, whether to use the absolute values of the statistic}
}
}
\description{
This function returns the default ranking options for the the
statistical measure in input.
}
\author{
Alessandro Barberis
}
\keyword{internal}
