% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-featscreen-class.R
\name{getFeatureDimensionality}
\alias{getFeatureDimensionality}
\title{Get the dimension of the feature space}
\usage{
getFeatureDimensionality(x)
}
\arguments{
\item{x}{an object of class \code{\link{featscreen}}.}
}
\value{
An integer value, the number of screened features.
}
\description{
Get the feature space dimension from a \code{\link{featscreen}} object.
}
\author{
Alessandro Barberis
}
