% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-featscreen-class.R
\name{getFeatureNames}
\alias{getFeatureNames}
\title{Get the feature names}
\usage{
getFeatureNames(x)
}
\arguments{
\item{x}{an object of class \code{\link{featscreen}}.}
}
\value{
A character vector indicating the feature names.
}
\description{
Get the feature names from a \code{\link{featscreen}} object.
}
\author{
Alessandro Barberis
}
