% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-featscreen-class.R
\name{getFeatureRanks}
\alias{getFeatureRanks}
\title{Get the feature ranks}
\usage{
getFeatureRanks(x)
}
\arguments{
\item{x}{an object of class \code{\link{featscreen}}.}
}
\value{
An integer vector indicating the ordering feature indices.
}
\description{
Get the feature ranks from a \code{\link{featscreen}} object.
}
\author{
Alessandro Barberis
}
