% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-featscreen-class.R
\name{getMultiresponseAggregationMethodId}
\alias{getMultiresponseAggregationMethodId}
\title{Get the multi-response aggregation method id}
\usage{
getMultiresponseAggregationMethodId(x)
}
\arguments{
\item{x}{an object of class \code{\link{featscreen}}.}
}
\value{
A character string, the multi-response aggregation method id.
}
\description{
Get the multi-response aggregation method id from a \code{\link{featscreen}} object.
}
\author{
Alessandro Barberis
}
