% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{getMultiresponseAggregationMethodName}
\alias{getMultiresponseAggregationMethodName}
\title{Multiresponse Aggregation Method Name}
\usage{
getMultiresponseAggregationMethodName(
  id = c("max", "min", "avg", "sum", "idx")
)
}
\arguments{
\item{id}{character string, one of the supported multi-response aggregation methods.}
}
\value{
A character string, the name of the multi-response aggregation method.
}
\description{
This function returns the name of the multi-response aggregation
method in input.
}
\author{
Alessandro Barberis
}
\keyword{internal}
