% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-featscreen-class.R
\name{getScreenedFeatures}
\alias{getScreenedFeatures}
\title{Get the screened features}
\usage{
getScreenedFeatures(x)
}
\arguments{
\item{x}{an object of class \code{\link{featscreen}}.}
}
\value{
A logical or integer vector indicating the features to keep.
If an integer vector of length equal to the dimension of the feature space is
returned, it is assumed to contain the indices of the ranked features.
}
\description{
Get the screened features from a \code{\link{featscreen}} object.
}
\author{
Alessandro Barberis
}
