% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{getScreeningMethodDataType}
\alias{getScreeningMethodDataType}
\title{Screening Method Data Type}
\usage{
getScreeningMethodDataType(
  id = c("cor.test", "pearson", "spearman", "kendall", "t.test", "t.test.equal",
    "t.test.unequal", "t.test.paired", "w.test", "w.test.ranksum", "w.test.paired",
    "anova", "anova.equal", "anova.unequal", "kruskal.wallis", "chisq.test", "coxph",
    "moderated.t", "moderated.F", "sam.test", "missing.value", "above.median",
    "above.minimum", "median", "variability")
)
}
\value{
A character string or a string vector, the data types supported by the
screening method.
}
\description{
This function returns the data type(s) the screening method in input
can work on.
}
\author{
Alessandro Barberis
}
\keyword{internal}
