% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-featscreen-class.R
\name{getScreeningMethodId}
\alias{getScreeningMethodId}
\title{Get the screening method id}
\usage{
getScreeningMethodId(x)
}
\arguments{
\item{x}{an object of class \code{\link{featscreen}}}
}
\value{
A character string, the screening method id.
}
\description{
Get the screening method id from a \code{\link{featscreen}} object.
}
\author{
Alessandro Barberis
}
