% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-selection-functions.R
\name{getSelectionFunction}
\alias{getSelectionFunction}
\title{Get Selection Function}
\usage{
getSelectionFunction(id = c("cutoff", "rank", "percentile", "fpr", "fdr"))
}
\arguments{
\item{id}{character string, one of the supported selecting techniques.}
}
\value{
A selecting function:

\describe{
\item{\code{"cutoff"    }}{\code{\link{selectByCutoff}}}
\item{\code{"rank"      }}{\code{\link{selectByRanking}}}
\item{\code{"percentile"}}{\code{\link{selectByPercentile}}}
\item{\code{"fpr"       }}{\code{\link{selectByFpr}}}
\item{\code{"fdr"       }}{\code{\link{selectByFdr}}}
}
}
\description{
This function is a dispatcher for the selecting function
in input.
}
\examples{
f = getSelectionFunction(id = 'cutoff')

}
\author{
Alessandro Barberis
}
