% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{getSelectionFunctionName}
\alias{getSelectionFunctionName}
\title{Selection Function Name}
\usage{
getSelectionFunctionName(id = c("cutoff", "rank", "percentile", "fpr", "fdr"))
}
\arguments{
\item{id}{character string, one of the supported selection functions.}
}
\value{
A character string, the name of the selection function.
}
\description{
This function returns the name of the selection function in input.
}
\author{
Alessandro Barberis
}
\keyword{internal}
