% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-featscreen-class.R
\name{getSelectionMethodId}
\alias{getSelectionMethodId}
\title{Get the selection method id}
\usage{
getSelectionMethodId(x)
}
\arguments{
\item{x}{an object of class \code{\link{featscreen}}.}
}
\value{
A character string, the selection method id.
}
\description{
Get the selection method id from a \code{\link{featscreen}} object.
}
\author{
Alessandro Barberis
}
