% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{getSelectionMethodName}
\alias{getSelectionMethodName}
\title{Selection Method Name}
\usage{
getSelectionMethodName(id = c("cutoff", "rank", "percentile", "fpr", "fdr"))
}
\arguments{
\item{id}{character string, one of the supported selecting methods.}
}
\value{
A character string, the name of the selecting method.
}
\description{
This function returns the name of the selection method in input.
}
\author{
Alessandro Barberis
}
\keyword{internal}
