% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-screen-function.R
\name{getStatFunction}
\alias{getStatFunction}
\title{Get Statistical Function}
\usage{
getStatFunction(
  id = c("cor.test", "pearson", "spearman", "kendall", "t.test", "t.test.equal",
    "t.test.unequal", "t.test.paired", "w.test", "w.test.ranksum", "w.test.paired",
    "anova", "anova.equal", "anova.unequal", "kruskal.wallis", "chisq.test", "coxph",
    "moderated.t", "moderated.F", "sam.test", "missing.value", "above.median",
    "above.minimum", "median", "variability")
)
}
\arguments{
\item{id}{character string, one of the supported statistical techniques.}
}
\value{
A statistical function:

\describe{
\item{\code{"cor.test"      }}{\code{\link{rowCor}}}
\item{\code{"pearson"       }}{\code{\link{rowPearsonCor}}}
\item{\code{"spearman"      }}{\code{\link{rowSpearmanCor}}}
\item{\code{"kendall"       }}{\code{\link{rowKendallCor}}}
\item{\code{"t.test"        }}{\code{\link{rowTwoSampleT}}}
\item{\code{"t.test.equal"  }}{\code{\link{rowEqualVarT}}}
\item{\code{"t.test.unequal"}}{\code{\link{rowUnequalVarT}}}
\item{\code{"t.test.paired" }}{\code{\link{rowPairedT}}}
\item{\code{"w.test"        }}{\code{\link{rowTwoSampleWilcoxonT}}}
\item{\code{"w.test.ranksum"}}{\code{\link{rowWilcoxonT}}}
\item{\code{"w.test.paired" }}{\code{\link{rowPairedWilcoxonT}}}
\item{\code{"anova"         }}{\code{\link{rowOneWayAnova}}}
\item{\code{"anova.equal"   }}{\code{\link{rowEqualVarOneWayAnova}}}
\item{\code{"anova.unequal" }}{\code{\link{rowUnequalVarOneWayAnova}}}
\item{\code{"kruskal.wallis"}}{\code{\link{rowKruskalWallis}}}
\item{\code{"chisq.test"    }}{\code{\link{rowPearsonChiSq}}}
\item{\code{"coxph"         }}{\code{\link{rowCoxPH}}}
\item{\code{"moderated.t"   }}{\code{\link{rowModeratedT}}}
\item{\code{"moderated.F"   }}{\code{\link{rowModeratedOneWayAnova}}}
\item{\code{"sam.test"      }}{\code{\link{rowSamStatistics}}}
\item{\code{"missing.value" }}{\code{\link{rowMissingValueRatio}}}
\item{\code{"above.median"  }}{\code{\link{rowAboveMedianFreqRatio}}}
\item{\code{"above.minimum" }}{\code{\link{rowAboveMinFreqRatio}}}
\item{\code{"median"        }}{\code{\link{rowMedians}}}
\item{\code{"variability"   }}{\code{\link{rowVariability}}}
}
}
\description{
This function is a dispatcher for the statistical function selected
in input.
}
\examples{
f = getStatFunction(id = 'cor.test')

}
\author{
Alessandro Barberis
}
