% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{getStatisticalMethodName}
\alias{getStatisticalMethodName}
\title{Statistical Method Name}
\usage{
getStatisticalMethodName(
  id = c("cor.test", "pearson", "spearman", "kendall", "t.test", "t.test.equal",
    "t.test.unequal", "t.test.paired", "w.test", "w.test.ranksum", "w.test.paired",
    "anova", "anova.equal", "anova.unequal", "kruskal.wallis", "chisq.test", "coxph",
    "moderated.t", "moderated.F", "sam.test", "missing.value", "above.median",
    "above.minimum", "median", "variability")
)
}
\arguments{
\item{id}{character string, one of the supported statistical techniques.}
}
\value{
A character string, the name of the statistical technique.
}
\description{
This function returns the name of the statistical method in input.
}
\author{
Alessandro Barberis
}
\keyword{internal}
