% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{getStatisticCategory}
\alias{getStatisticCategory}
\title{Screening Test Category}
\usage{
getStatisticCategory(
  id = c("cor.test", "pearson", "spearman", "kendall", "t.test", "t.test.equal",
    "t.test.unequal", "t.test.paired", "w.test.ranksum", "w.test.paired", "anova",
    "anova.equal", "anova.unequal", "kruskal.wallis", "chisq.test", "coxph",
    "moderated.t", "moderated.F", "sam.test", "missing.value", "above.median",
    "above.minimum", "median", "variability")
)
}
\arguments{
\item{id}{character string, one of the supported statistical techniques.}
}
\value{
A character string, the name of the assigned group.
}
\description{
This function returns the category of the statistic used for
screening.
}
\author{
Alessandro Barberis
}
\keyword{internal}
