% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-featscreen-class.R
\name{getSummary}
\alias{getSummary}
\title{Get the screening summary}
\usage{
getSummary(x)
}
\arguments{
\item{x}{an object of class \code{\link{featscreen}}.}
}
\value{
A character string, the textual summary of the screening process.
}
\description{
Get the summary from a \code{\link{featscreen}} object.
}
\author{
Alessandro Barberis
}
