% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{getVariabilityMeasureName}
\alias{getVariabilityMeasureName}
\title{Variability Measure Name}
\usage{
getVariabilityMeasureName(
  id = c("sd", "iqr", "mad", "rsd", "efficiency", "vmr")
)
}
\arguments{
\item{id}{character string, one of the supported variability measures.}
}
\value{
A character string, the name of the variability measure.
}
\description{
This function returns the name of the variability measure in input.
}
\author{
Alessandro Barberis
}
\keyword{internal}
