% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-stats-functions.R
\name{iqr}
\alias{iqr}
\title{Interquartile Range}
\usage{
iqr(x, g = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{numerical vector.}

\item{g}{(optional) vector or factor object giving the group for the corresponding
elements of \code{x}.}

\item{na.rm}{logical indicating whether missing values should be removed before
computation.}
}
\description{
This function computes the \emph{interquartile range} (IQR) of the values in
\code{x}.

See the \strong{Details} section below for further information.
}
\details{
This function is a wrapper to the \code{\link[stats]{IQR}} function.

If \code{x} can be partitioned into \eqn{c} subgroups (provided by \code{g}),
then the \eqn{IQR} is computed for each class.
}
\examples{
#Seed
set.seed(1010)

#Define size
n = 10

#Data
x = sample.int(n = 100, size = n, replace = TRUE)

#Grouping variable
g = c(rep("a", n/2), rep("b", n/2))

#Interquartile Range
iqr(x)

#Interquartile Range by group
iqr(x = x, g = g)

}
