% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-featscreen-class.R
\name{is.featscreen}
\alias{is.featscreen}
\title{Check a \code{featscreen} object}
\usage{
is.featscreen(x)
}
\arguments{
\item{x}{an object of class \code{\link{featscreen}}.}
}
\value{
Returns \code{TRUE} if its argument is a valid \code{\link{featscreen}} object.
}
\description{
Function to check if an object is a \code{\link{featscreen}} object.
}
\examples{
#valid object
is.featscreen(x = featscreen())

#invalid object
is.featscreen(x = 3)

}
\author{
Alessandro Barberis
}
