% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-log-functions.R
\name{is.Logger}
\alias{is.Logger}
\title{Check class}
\usage{
is.Logger(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
Logical, whether \code{x} is of class \code{\link{Logger}}.
}
\description{
Function to check if input object is of class \code{\link{Logger}}.
}
\author{
Alessandro Barberis
}
\keyword{internal}
