% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{listAvailableScreeningMethodIds}
\alias{listAvailableScreeningMethodIds}
\title{Available Screening Method IDs}
\usage{
listAvailableScreeningMethodIds(
  x = c("all", "unsupervised", "supervised"),
  show.wrapper = FALSE
)
}
\arguments{
\item{x}{character string indicating whether to return all screening methods
(\code{all}), the unsupervised screening methods (\code{unsupervised}) or the supervised
screening methods (\code{supervised}).}

\item{show.wrapper}{logical indicating whether to return the ids for the
supervised screening functions with or without wrapper functions.}
}
\value{
A vector containing the screening method identifiers.
}
\description{
This function returns the currently available
random sampling methods.
}
\author{
Alessandro Barberis
}
\keyword{internal}
