% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{listAvailableScreeningMethods}
\alias{listAvailableScreeningMethods}
\title{Available Screening Methods}
\usage{
listAvailableScreeningMethods(
  x = c("all", "unsupervised", "supervised"),
  show.wrapper = FALSE,
  show.data.type = FALSE,
  simplify = FALSE,
  show.function = FALSE,
  show.category = FALSE
)
}
\arguments{
\item{x}{character string indicating whether to return all screening methods
(\code{all}), the unsupervised screening methods (\code{unsupervised}) or the supervised
screening methods (\code{supervised}).}

\item{show.wrapper}{logical indicating whether to return the ids for the
supervised screening functions with or without wrapper functions.}

\item{show.data.type}{logical indicating whether to show the supported data types.}

\item{simplify}{logical indicating whether to format the data for readability.}

\item{show.function}{logical indicating whether to show the function name.}

\item{show.category}{logical indicating whether to show the internally given category.}
}
\value{
A data frame with at least two columns:

\describe{
\item{id}{the id of the screening method, to be used in the function calls}
\item{name}{the name of the screening method}
}
}
\description{
This function returns the currently available
screening methods.
}
\examples{

#List available methods
listAvailableScreeningMethods()

#List available methods and supported data types
listAvailableScreeningMethods(show.data.type = TRUE, simplify = TRUE)

#List available methods and function names
listAvailableScreeningMethods(show.function = TRUE)

}
\author{
Alessandro Barberis
}
