% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{listAvailableSelectionFunctions}
\alias{listAvailableSelectionFunctions}
\title{Available Selection Functions}
\usage{
listAvailableSelectionFunctions()
}
\value{
A data frame with two columns:

\describe{
\item{id}{the id of the selection method, to be used in the function calls}
\item{name}{the name of the selection function}
}
}
\description{
This function returns the currently available
selection function names.
}
\examples{

#List available methods
listAvailableSelectionFunctions()

}
\author{
Alessandro Barberis
}
