% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{listAvailableSelectionMethodIds}
\alias{listAvailableSelectionMethodIds}
\title{Available Selection Method Ids}
\usage{
listAvailableSelectionMethodIds()
}
\value{
A vector containing the selecting method identifiers.
}
\description{
This function returns the ids of the currently available methods
for feature selection.
}
\author{
Alessandro Barberis
}
\keyword{internal}
