% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{listAvailableSelectionMethods}
\alias{listAvailableSelectionMethods}
\title{Available Selection Methods}
\usage{
listAvailableSelectionMethods(show.function = FALSE)
}
\arguments{
\item{show.function}{logical indicating whether to show the related function name.}
}
\value{
A data frame with two columns by default:

\describe{
\item{id}{the id of the selection method, to be used in the function calls}
\item{name}{the name of the selection method}
}

If \code{show.function = TRUE}, a further column containing the function name
is added:

\describe{
\item{id}{the id of the selection method, to be used in the function calls}
\item{name}{the name of the selection method}
\item{function}{the name of the selection function}
}
}
\description{
This function returns the currently available
selection methods.
}
\examples{

#List available methods
listAvailableSelectionMethods()

#List available methods and related functions
listAvailableSelectionMethods(show.function = TRUE)

}
\author{
Alessandro Barberis
}
