% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{listAvailableSupervisedScreeningMethodIds}
\alias{listAvailableSupervisedScreeningMethodIds}
\title{Available Supervised Screening Method Ids}
\usage{
listAvailableSupervisedScreeningMethodIds(show.wrapper = FALSE)
}
\arguments{
\item{show.wrapper}{logical indicating whether to return the ids for the
supervised screening functions with or without wrapper functions.}
}
\value{
A vector containing the screening method identifiers.
}
\description{
This function returns the ids of the currently available methods
for supervised screening.
}
\author{
Alessandro Barberis
}
\keyword{internal}
