% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{listAvailableUnsupervisedScreeningMethodIds}
\alias{listAvailableUnsupervisedScreeningMethodIds}
\title{Available Unsupervised Screening Method Ids}
\usage{
listAvailableUnsupervisedScreeningMethodIds()
}
\value{
A vector containing the screening method identifiers.
}
\description{
This function returns the ids of the currently available methods
for unsupervised screening.
}
\author{
Alessandro Barberis
}
\keyword{internal}
