% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{listDefaultRankingOption}
\alias{listDefaultRankingOption}
\title{Default Ranking Options}
\usage{
listDefaultRankingOption(
  id = c("cor.test", "pearson", "spearman", "kendall", "t.test", "t.test.equal",
    "t.test.unequal", "t.test.paired", "w.test", "w.test.ranksum", "w.test.paired",
    "anova", "anova.equal", "anova.unequal", "kruskal.wallis", "chisq.test", "coxph",
    "moderated.t", "moderated.F", "sam.test", "missing.value", "above.median",
    "above.minimum", "median", "variability")
)
}
\value{
A data frame with five columns:

\describe{
\item{id}{the id of the screening method, to be used in the function calls}
\item{statistic}{the statistic associated with the screening method}
\item{measure}{the measure to use for ranking, whether to use the statistic or the significance (where computed)}
\item{order}{whether to sort the measure in ascending or descending order for ranking}
\item{value}{whether to use absolute values for ranking}
}
}
\description{
This function returns the default ranking options for the statistics
of the screening method in input.
}
\author{
Alessandro Barberis
}
\keyword{internal}
