% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{listDefaultRankingOptions}
\alias{listDefaultRankingOptions}
\title{Default Ranking Options}
\usage{
listDefaultRankingOptions()
}
\value{
A data frame with five columns:

\describe{
\item{id}{the id of the screening method, to be used in the function calls}
\item{statistic}{the statistic associated with the screening method}
\item{measure}{the measure to use for ranking, whether to use the statistic or the significance (where computed)}
\item{order}{whether to sort the measure in ascending or descending order for ranking}
\item{value}{whether to use absolute values for ranking}
}
}
\description{
This function returns the default ranking options for the statistics
of the screening methods.
}
\examples{

#List available methods
listDefaultRankingOptions()

}
\author{
Alessandro Barberis
}
