% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{listScreeningMethodCategories}
\alias{listScreeningMethodCategories}
\title{Available Screening Functions}
\usage{
listScreeningMethodCategories(
  x = c("all", "unsupervised", "supervised"),
  show.wrapper = FALSE
)
}
\arguments{
\item{x}{character string indicating whether to return all screening methods
(\code{all}), the unsupervised screening methods (\code{unsupervised}) or the supervised
screening methods (\code{supervised}).}

\item{show.wrapper}{logical indicating whether to return the ids for the
supervised screening functions with or without wrapper functions.}
}
\value{
A data frame with two columns:

\describe{
\item{id}{the id of the screening function, to be used in the function calls}
\item{category}{the name of the internally given category}
}
}
\description{
This function returns the internally given category for each
available screening method.
}
\author{
Alessandro Barberis
}
\keyword{internal}
