% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{listScreeningMethodDataTypes}
\alias{listScreeningMethodDataTypes}
\title{Available Screening Methods}
\usage{
listScreeningMethodDataTypes(
  x = c("all", "unsupervised", "supervised"),
  show.wrapper = FALSE,
  simplify = FALSE
)
}
\arguments{
\item{x}{character string indicating whether to return all screening methods
(\code{all}), the unsupervised screening methods (\code{unsupervised}) or the supervised
screening methods (\code{supervised}).}

\item{show.wrapper}{logical indicating whether to return the ids for the
supervised screening functions with or without wrapper functions.}

\item{simplify}{logical indicating whether to format the data for readability.}
}
\value{
A data frame with five columns:

\describe{
\item{id}{the id of the screening method, to be used in the function calls}
\item{numerical}{whether the method works with numerical data}
\item{categorical}{whether the method works with categorical data}
\item{multiresponse}{whether the method works with multiresponse data}
\item{survival}{whether the method works with survival data}
}
}
\description{
This function returns the supported data types for each
available screening method.
}
\details{
If \code{simplify = FALSE}, the values of the table are returned as
logical (i.e., TRUE and FALSE).
If \code{simplify = TRUE}, the values of the table are returned as 'x' (meaning
the method works with the indicated data type) and ''.
}
\author{
Alessandro Barberis
}
\keyword{internal}
