% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{multiresponse}
\alias{multiresponse}
\title{Combine multi-response}
\usage{
multiresponse(x, multi = c("max", "min", "avg", "sum", "idx"), idx = NULL)
}
\arguments{
\item{x}{\code{matrix} containing the values to combine.}

\item{multi}{character string indicating what to do when \code{x} has multiple
columns.

Available options are:
\describe{
\item{\code{"max"}}{the maximum value across responses is kept}
\item{\code{"min"}}{the minimum value across responses is kept}
\item{\code{"avg"}}{values are averaged}
\item{\code{"sum"}}{values are summed up}
\item{\code{"idx"}}{return the column indicated by \code{idx}}
}}

\item{idx}{(optional) integer value or character string indicating the
column of \code{x} to keep.}
}
\value{
A vector.
}
\description{
If screened features are from multi-response data,
this function allows to adopt different strategies.
}
\author{
Alessandro Barberis
}
