% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-featscreen-class.R
\name{print.featscreen}
\alias{print.featscreen}
\title{Print a \code{featscreen} object}
\usage{
\method{print}{featscreen}(x, show.top = TRUE, top = 5L, show.names = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{\link{featscreen}}.}

\item{show.top}{logical indicating whether to show \code{top} ranked features.}

\item{top}{integer, the number of top ranked features to show.}

\item{show.names}{logical indicating whether to show indices or names.}

\item{...}{additional print arguments}
}
\value{
Silently return \code{x}.
}
\description{
Print a summary of the \code{\link{featscreen}} object.
}
\examples{
#featscreen object
x = featscreen(
 method = 'cor.test',
 selection = 'cutoff',
 summary = '3 out of 5 features selected by a cutoff.',
 n = 5,
 keep = c(TRUE,FALSE,TRUE,FALSE,TRUE),
 features = paste0("f", seq_len(5)),
 ranks = c(1,5,3,4,2)
)

#print
print(x)

}
