% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-stats-functions.R
\name{rowAboveMedianFreqRatio}
\alias{rowAboveMedianFreqRatio}
\title{Above-Median Frequency Ratio}
\usage{
rowAboveMedianFreqRatio(x, g = NULL)
}
\arguments{
\item{x}{\code{matrix} or \code{data.frame}, where rows are features and columns are observations.}

\item{g}{(optional) vector or factor object giving the group for the corresponding
elements of \code{x}.}
}
\value{
A \code{vector} of length \code{nrow(x)} containing the computed ratios.
If \code{g} is provided, a \code{matrix} with ratios for each class as column
vectors is returned.
}
\description{
This function computes the ratio of values above the median for
each feature.

See the \strong{Details} section below for further information.
}
\details{
For each observation, the median is computed via
\code{\link[matrixStats]{colMedians}}.

Remember that the median across \eqn{n} elements is defined as:

\deqn{Median = x_\frac{(n+1)}{2}}

where \eqn{x} is an ascendingly ordered vector of \eqn{n} elements, and \eqn{n}
is odd. If \eqn{n} is even, then the median is computed as:

\deqn{Median = \frac{x_\frac{n}{2}+x_{\frac{(n)}{2}+1}}{2}}

If \code{g = NULL}, for each feature we define an above-median frequency ratio (AMFR) as the
number of times the feature value is greater than the sample median divided
by the total number of observations:

\deqn{Above-Median Frequency Ratio (AMFR) = \frac{Number of samples where feature is above the sample median}{Total number of observations}}

If \code{g} is provided, the above-median frequency ratio (\eqn{AMFR_{ij}}) is
computed for each group \eqn{j} as:

\deqn{AMFR_{ij} = \frac{Number of samples in j-th class where feature is above the sample median}{Number of observations in j-th class}}
}
\examples{
#Seed
set.seed(1010)

#Define row/col size
nr = 5
nc = 10

#Data
x = matrix(
 data = sample.int(n = 100, size = nr*nc, replace = TRUE),
 nrow = nr,
 ncol = nc,
 dimnames = list(
   paste0("f",seq(nr)),
   paste0("S",seq(nc))
 )
)

#Grouping variable
g = c(rep("a", nc/2), rep("b", nc/2))

#AMR
rowAboveMedianFreqRatio(x)

#AMR by group
rowAboveMedianFreqRatio(x = x, g = g)

}
\author{
Alessandro Barberis
}
