% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-stats-functions.R
\name{rowAboveMinFreqRatio}
\alias{rowAboveMinFreqRatio}
\title{Above-Minimum Frequency Ratio}
\usage{
rowAboveMinFreqRatio(x, g = NULL, min.expr = 0)
}
\arguments{
\item{x}{\code{matrix} or \code{data.frame}, where rows are features and columns are observations.}

\item{g}{(optional) vector or factor object giving the group for the corresponding
elements of \code{x}.}

\item{min.expr}{numerical value indicating the minimum expression required for
\code{min.prop} samples.}
}
\value{
A \code{vector} of length \code{nrow(x)} containing the computed ratios.
If \code{g} is provided, a \code{matrix} with ratios for each class as column
vectors is returned.
}
\description{
This function computes the ratio of values above a minimum value for
each feature.

See the \strong{Details} section below for further information.
}
\details{
For each feature, the above-minimum frequency ratio (AMFR) is computed as:

\deqn{Above-Minimum Frequency Ratio (AMFR) = \frac{Number of samples where expression is greater than provided minimum}{Total number of observations}}

If \code{g} is provided, the above-minimum frequency ratio (\eqn{AMFR_{ij}}) is
computed for each group \eqn{j} as:

\deqn{AMFR_{ij} = \frac{Number of samples in j-th class where expression is greater than provided minimum}{Total number of observations in j-th class}}
}
\examples{
#Seed
set.seed(1010)

#Define row/col size
nr = 5
nc = 10

#Data
x = matrix(
 data = sample.int(n = 100, size = nr*nc, replace = TRUE),
 nrow = nr,
 ncol = nc,
 dimnames = list(
   paste0("f",seq(nr)),
   paste0("S",seq(nc))
 )
)

#Grouping variable
g = c(rep("a", nc/2), rep("b", nc/2))

#AMFR
rowAboveMinFreqRatio(x)

#AMFR by group
rowAboveMinFreqRatio(x = x, g = g)

}
\author{
Alessandro Barberis
}
