% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-test-functions.R
\name{rowCor}
\alias{rowCor}
\title{Row Correlations}
\usage{
rowCor(
  x,
  y,
  alternative = c("two.sided", "greater", "less"),
  method = c("pearson", "kendall", "spearman"),
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{x}{\code{matrix} or \code{data.frame}.}

\item{y}{numeric vector of data values. Must have the same length as \code{ncol(x)}.}

\item{alternative}{character string or vector of length \code{nrow(x)}.
The alternative hypothesis for each row of \code{x}.
Values must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"less"}.}

\item{method}{character string indicating how to measure the association.
Available options are:
\describe{
\item{\code{"pearson"}}{Pearson's product moment coefficient}
\item{\code{"kendall"}}{Kendall's tau}
\item{\code{"spearman"}}{Spearman's rho}
}}

\item{conf.level}{numerical value or numeric vector of length \code{nrow(x)}.
The confidence levels of the intervals.
All values must be in the range of \eqn{[0:1]} or \code{NA}.}

\item{...}{further arguments to \code{\link[stats]{cor.test}}}
}
\value{
A list containing two elements:

\describe{
\item{statistic}{A numeric vector, the values of the test statistic}
\item{significance}{A numeric vector, the p-values of the selected test}
}
}
\description{
This function tests for association between each row vector in
\code{x} and \code{y}, using one of Pearson's product moment coefficient,
Kendall's tau or Spearman's rho.

See the \strong{Details} section below for further information.
}
\details{
See the following functions for each specific implementation:

\describe{
\item{\code{"pearson" }}{\code{\link{rowPearsonCor}}}
\item{\code{"kendall" }}{\code{\link{rowKendallCor}}}
\item{\code{"spearman"}}{\code{\link{rowSpearmanCor}}}
}
}
\examples{
#Seed
set.seed(1010)

#Data
x = matrix(rnorm(100 * 20), 100, 20)
y = rnorm(20)

#Compute
rowCor(x = x, y = y)

}
\author{
Alessandro Barberis
}
