% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-test-functions.R
\name{rowCoxPH}
\alias{rowCoxPH}
\title{Fit Univariate Proportional Hazards Regression Model}
\usage{
rowCoxPH(x, y, ...)
}
\arguments{
\item{x}{independent variable}

\item{y}{\code{data.frame} with two columns, \code{time} and \code{status}}

\item{...}{further arguments to \code{\link[survival]{coxph}}}
}
\value{
A list containing two elements:

\describe{
\item{statistic}{A numeric vector, the values of the test statistic}
\item{significance}{A numeric vector, the p-values of the selected test}
}
}
\description{
Fits a Cox proportional hazards regression model for each
row vector in \code{x}.
}
\details{
It is a wrapper to \code{\link[survival]{coxph}}
function.
}
\examples{
#Seed
set.seed(1010)

#Data
x = matrix(rnorm(100 * 7), 100, 7)
y = data.frame(
  time = c(4,3,1,1,2,2,3),
  status = c(1,1,1,0,1,1,0)
)

#Compute
rowCoxPH(x = x, y = y)

}
\author{
Alessandro Barberis
}
