% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-test-functions.R
\name{rowEqualVarOneWayAnova}
\alias{rowEqualVarOneWayAnova}
\title{Fit an Analysis of Variance Model}
\usage{
rowEqualVarOneWayAnova(x, g)
}
\arguments{
\item{x}{numeric matrix.}

\item{g}{a vector specifying group membership for each observation of x.}
}
\value{
A list containing two elements:

\describe{
\item{statistic}{A numeric vector, the values of the test statistic}
\item{significance}{A numeric vector, the p-values of the selected test}
}
}
\description{
Computes the one-way analysis of variance test for each feature.

See the \strong{Details} section below for further information.
}
\details{
It is a wrapper to \code{\link[matrixTests]{row_oneway_equalvar}}
function.
}
\examples{
#Seed
set.seed(1010)

#Data
x = matrix(rnorm(100 * 20), 100, 20)
g = sample(c(0,1), 20, replace = TRUE)

#Compute
rowEqualVarOneWayAnova(x = x, g = g)

}
\author{
Alessandro Barberis
}
