% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-filter-functions.R
\name{rowFilterByAboveMinRatio}
\alias{rowFilterByAboveMinRatio}
\title{Filter by Above-Minimum Frequency Ratio}
\usage{
rowFilterByAboveMinRatio(x, g = NULL, min.expr = 0, min.prop = 0.5)
}
\arguments{
\item{x}{\code{matrix} or \code{data.frame}, where rows are features and columns are observations.}

\item{g}{(optional) vector or factor object giving the group for the corresponding
elements of \code{x}.}

\item{min.expr}{numerical value indicating the minimum expression required for
\code{min.prop} samples.}

\item{min.prop}{numerical value in the range \eqn{[0, 1]}. Minimum proportion of
samples where the feature expression should be above \code{min.expr}.
Default to \code{0.5}.}
}
\value{
A logical vector of length \code{nrow(x)} indicating which rows of
\code{x} passed the filter.
}
\description{
This function filters the input matrix \code{x} depending on the
features' intensity.
Variables are removed if their values are lower than a provided minimum value
in a given percentage of samples defined by \code{min.prop}.

See the \strong{Details} section below for further information.
}
\details{
For each feature, the above-minimum frequency ratio (AMFR) is computed as:

\deqn{Above-Minimum Frequency Ratio (AMFR) = \frac{Number of samples where expression is greater than provided minimum}{Total number of observations}}

Then, the i-th feature is kept if \eqn{AMFR_{i} >= min.prop}.

If \code{g} is provided, the above-minimum frequency ratio (\eqn{AMFR_{ij}}) is
computed for each group \eqn{j}.
The i-th feature is kept if \eqn{AMFR_{ij} >= min.prop} for at least one group.
}
\examples{
#Seed
set.seed(1010)

#Define row/col size
nr = 5
nc = 10

#Data
x = matrix(
 data = sample.int(n = 100, size = nr*nc, replace = TRUE),
 nrow = nr,
 ncol = nc,
 dimnames = list(
   paste0("f",seq(nr)),
   paste0("S",seq(nc))
 )
)

#Grouping variable
g = c(rep("a", nc/2), rep("b", nc/2))

#Filter
rowFilterByAboveMinRatio(x)

#Filter by group
rowFilterByAboveMinRatio(x = x, g = g)

#Set 1st feature to 0s for 2/3 observations
x[1,seq(2*nc/3)] = 0

#Set 3rd feature to 0s for 2/3 observations of class "a" and "b"
x[3,seq(2*nc/6)] = 0
x[3,(seq(2*nc/6)+nc/2)] = 0

#Filter (1st and 3rd features should be flagged to be removed)
rowFilterByAboveMinRatio(x = x, min.expr = 10)

#Filter by group (3rd feature should be flagged to be removed)
rowFilterByAboveMinRatio(x = x, min.expr = 10, g = g)

}
\author{
Alessandro Barberis
}
