% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-filter-functions.R
\name{rowFilterByLowVariability}
\alias{rowFilterByLowVariability}
\title{Filter by Low Variability}
\usage{
rowFilterByLowVariability(
  x,
  g = NULL,
  percentile = 0.25,
  method = c("sd", "iqr", "mad", "rsd", "vmr", "efficiency")
)
}
\arguments{
\item{x}{\code{matrix} or \code{data.frame}, where rows are features and columns are observations.}

\item{g}{(optional) vector or factor object giving the group for the corresponding
elements of \code{x}.}

\item{percentile}{numerical value in the range \eqn{[0, 1]} indicating the
percentage of features to keep.}

\item{method}{character string indicating the measure of variability.
Available options are:
\describe{
\item{\code{"sd"}}{the standard deviation}
\item{\code{"iqr"}}{the interquartile range}
\item{\code{"mad"}}{the median absolute deviation}
\item{\code{"rsd"}}{the relative standard deviation (i.e., coefficient of variation)}
\item{\code{"efficiency"}}{the coefficient of variation squared}
\item{\code{"vmr"}}{the variance-to-mean ratio}
}}
}
\value{
A logical vector of length \code{nrow(x)} indicating which rows of
\code{x} passed the filter.
}
\description{
This function filters the input matrix \code{x} depending on the
features' variability.

See the \code{\link{rowVariability}} for further information.
}
\examples{
#Seed
set.seed(1010)

#Define row/col size
nr = 5
nc = 10

#Data
x = matrix(
 data = sample.int(n = 100, size = nr*nc, replace = TRUE),
 nrow = nr,
 ncol = nc,
 dimnames = list(
   paste0("f",seq(nr)),
   paste0("S",seq(nc))
 )
)

#Grouping variable
g = c(rep("a", nc/2), rep("b", nc/2))

#Filter
rowFilterByLowVariability(x)

#Filter by group
rowFilterByLowVariability(x = x, g = g)

}
\author{
Alessandro Barberis
}
