% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-test-functions.R
\name{rowKruskalWallis}
\alias{rowKruskalWallis}
\title{Kruskal-Wallis Rank Sum Test}
\usage{
rowKruskalWallis(x, g)
}
\arguments{
\item{x}{\code{matrix} or \code{data.frame}.}

\item{g}{a vector or factor object giving the group for the corresponding
elements of \code{x}.}
}
\value{
A list containing two elements:

\describe{
\item{statistic}{A numeric vector, the values of the test statistic}
\item{significance}{A numeric vector, the p-values of the selected test}
}
}
\description{
Computes the Kruskal-Wallis rank sum test for each feature.

See the \strong{Details} section below for further information.
}
\details{
It is a wrapper to \code{\link[matrixTests]{row_kruskalwallis}}
function.
}
\examples{
#Seed
set.seed(1010)

#Data
x = matrix(rnorm(100 * 20), 100, 20)
g = sample(c(0,1), 20, replace = TRUE)

#Compute
rowKruskalWallis(x = x, g = g)

}
\author{
Alessandro Barberis
}
