% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-stats-functions.R
\name{rowMedians}
\alias{rowMedians}
\title{Medians}
\usage{
rowMedians(x, g = NULL)
}
\arguments{
\item{x}{\code{matrix} or \code{data.frame}, where rows are features and columns are observations.}

\item{g}{(optional) vector or factor object giving the group for the corresponding
elements of \code{x}.}
}
\value{
A \code{vector} of length \code{nrow(x)} containing the computed ratios.
If \code{g} is provided, a \code{matrix} with ratios for each class as column
vectors is returned.
}
\description{
This function computes the median value for
each feature.

See the \strong{Details} section below for further information.
}
\details{
If \code{g = NULL}, for each feature the median is computed via
\code{\link[matrixStats]{rowMedians}}.

If \code{g} is provided, the median per group is computed via
\code{\link[stats]{median}}.
}
\examples{
#Seed
set.seed(1010)

#Define row/col size
nr = 5
nc = 10

#Data
x = matrix(
 data = sample.int(n = 100, size = nr*nc, replace = TRUE),
 nrow = nr,
 ncol = nc,
 dimnames = list(
   paste0("f",seq(nr)),
   paste0("S",seq(nc))
 )
)

#Grouping variable
g = c(rep("a", nc/2), rep("b", nc/2))

#Medians
rowMedians(x)

#Medians by group
rowMedians(x = x, g = g)

}
\author{
Alessandro Barberis
}
