% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-stats-functions.R
\name{rowMissingValueRatio}
\alias{rowMissingValueRatio}
\title{Missing Value Ratio}
\usage{
rowMissingValueRatio(x, g = NULL)
}
\arguments{
\item{x}{\code{matrix} or \code{data.frame}, where rows are features and columns are observations.}

\item{g}{(optional) vector or factor object giving the group for the corresponding
elements of \code{x}.}
}
\value{
A \code{vector} of length \code{nrow(x)} containing the computed ratios.
If \code{g} is provided, a \code{matrix} with ratios for each class as column
vectors is returned.
}
\description{
This function computes the ratio of missing elements.

See the \strong{Details} section below for further information.
}
\details{
If \code{g = NULL}, for each feature a missing value ratio (MVR) is computed as:

\deqn{Missing Value Ratio (MVR) = \frac{Number of missing values}{Total number of observations}}

If \code{g} is provided, the missing value ratios \eqn{MVR_{ij}} are computed
for each group \eqn{j} as:

\deqn{MVR_{ij} = \frac{Number of missing values in j-th class}{Number of observations in j-th class}}
}
\examples{
#Seed
set.seed(1010)

#Define row/col size
nr = 5
nc = 10

#Data
x = matrix(
 data = sample(x = c(1,2), size = nr*nc, replace = TRUE),
 nrow = nr,
 ncol = nc,
 dimnames = list(
   paste0("f",seq(nr)),
   paste0("S",seq(nc))
 )
)

#Grouping variable
g = c(rep("a", nc/2), rep("b", nc/2))

#Force 1st feature to have 40\% of missing values
x[1,seq(nc*0.4)] = NA

#Compute MVR
rowMissingValueRatio(x = x)

#Compute MVR by class
rowMissingValueRatio(x = x, g = g)

}
\author{
Alessandro Barberis
}
