% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-biostats-test-functions.R
\name{rowSamStatistics}
\alias{rowSamStatistics}
\title{Significance Analysis of Microarrays}
\usage{
rowSamStatistics(
  x,
  y = NULL,
  observations = NULL,
  technology = c("array", "seq"),
  geneid = NULL,
  genenames = NULL,
  censoring.status = NULL,
  logged2 = FALSE,
  eigengene.number = 1,
  resp.type = c("Quantitative", "Two class unpaired", "Survival", "Multiclass",
    "One class", "Two class paired", "Two class unpaired timecourse",
    "One class timecourse", "Two class paired timecourse", "Pattern discovery"),
  s0 = NULL,
  s0.perc = NULL,
  nperms = 100,
  center.arrays = FALSE,
  testStatistic = c("standard", "wilcoxon"),
  time.summary.type = c("slope", "signed.area"),
  regression.method = c("standard", "ranks"),
  knn.neighbors = 10,
  random.seed = NULL,
  nresamp = 20,
  nresamp.perm = NULL,
  dels = NULL,
  nvals = 50,
  fdr.output = 0.2,
  logger = NULL
)
}
\arguments{
\item{x}{Feature matrix: p (number of features) by n (number of samples),
one observation per column (missing values allowed)}

\item{y}{n-vector of outcome measurements}

\item{observations}{(optional) integer vector, the indices of observations to
keep.}

\item{technology}{character string, the technology used to generate the data.
Available options are:
\describe{
\item{array}{data generated with microarray technology}
\item{seq}{data generated with RNA-seq technology}
}}

\item{geneid}{Optional character vector of geneids for output.}

\item{genenames}{Optional character vector of genenames for output.}

\item{censoring.status}{n-vector of censoring censoring.status (1= died or event occurred, 0=survived, or event was censored), needed for a censored survival outcome}

\item{logged2}{Has the data been transformed by log (base 2)? This information is used only 
for computing fold changes }

\item{eigengene.number}{Eigengene to be used  (just for resp.type="Pattern discovery")}

\item{resp.type}{Problem type:
"Quantitative" for a continuous parameter (Available for both array and sequencing data);
"Two class unpaired" (for both array and sequencing data); 
"Survival" for censored survival outcome (for both array and sequencing data); 
"Multiclass": more than 2 groups (for both array and sequencing data);  
"One class" for a single group (only for array data); 
"Two class paired" for two classes with paired observations (for both array and sequencing data);
"Two class unpaired timecourse" (only for array data), 
"One class time course" (only for array data),
"Two class.paired timecourse" (only for array data),
or "Pattern discovery" (only for array data)}

\item{s0}{Exchangeability factor  for denominator of test statistic; Default
is automatic choice. Only used for array data.}

\item{s0.perc}{Percentile of standard  deviation values to use for s0; default is
automatic choice; -1 means s0=0 (different from s0.perc=0, meaning
s0=zeroeth percentile of  standard  deviation values= min of sd values.
Only used for array data.}

\item{nperms}{Number of permutations used to estimate false discovery rates}

\item{center.arrays}{Should the data for each sample (array) be median centered
at the outset? Default =FALSE. Only used for array data.}

\item{testStatistic}{Test statistic to use in two class unpaired case.Either
"standard" (t-statistic) or ,"wilcoxon" (Two-sample wilcoxon or Mann-Whitney
test). Only used for array data.}

\item{time.summary.type}{Summary measure for each time course: "slope", or "signed.area"). Only used for array data.}

\item{regression.method}{Regression method for quantitative case:
"standard", (linear least squares) or "ranks" (linear least squares 
on ranked data). Only used for array data.}

\item{knn.neighbors}{Number of nearest neighbors to use for imputation
of missing features values. Only used for array data.}

\item{random.seed}{Optional initial seed for random number generator (integer)}

\item{nresamp}{For assay.type="seq", number of resamples used to construct test statistic. Default 20. Only used for sequencing data.}

\item{nresamp.perm}{For assay.type="seq", number of resamples used to construct test statistic for permutations. Default is equal to nresamp and it must be at most nresamp. Only used for sequencing data.}

\item{dels}{vector of delta values used. 
Delta is the vertical distance from the 45 degree line to the upper and lower
parallel lines that define the SAM threshold rule.
By default, for array data, 50 values
are chosen in the relevant operating change for delta.
For sequencing data, the maximum number of effective delta values
are chosen automatically according to the data.
}

\item{nvals}{Number of delta values used. 
For array data, the default value is 50. For sequencing data,
the value will be chosen automatically.}

\item{fdr.output}{(Approximate) False Discovery Rate cutoff for output in significant genes table}

\item{logger}{a \code{\link{Logger}} object.}
}
\value{
A list containing two elements:

\describe{
\item{statistic}{A numeric vector, the values of the test statistic}
\item{significance}{A numeric vector, the q-values of the selected test}
}
}
\description{
This function implements a common workflow to compute repeated
permutations of the input data to determine if any features are significantly
related to the response.

The following steps are executed:

\enumerate{
\item correlate each feature with outcome variable: \code{\link[samr]{samr}}
\item compute thresholds, cutpoints, and false discovery rates for SAM analysis: \code{\link[samr]{samr.compute.delta.table}}
\item compute SAM statistics and significance: \code{\link[samr]{samr.compute.siggenes.table}}
}
}
\examples{
#Seed
set.seed(1010)

#Define row/col size
nr = 10
nc = 20

#Data
x = matrix(
 data = stats::rnorm(n = nr*nc),
 nrow = nr,
 ncol = nc,
 dimnames = list(
   paste0("f",seq(nr)),
   paste0("S",seq(nc))
 )
)

#Categorical output vector (binomial)
y = c(rep(1,nc/2), rep(2,nc/2))
names(y) = paste0("S",seq(nc))
rowSamStatistics(x=x,y=y)

#Categorical output vector (multinomial)
y = c(rep(1,nc/4), rep(2,nc/4), rep(3,nc/2))
names(y) = paste0("S",seq(nc))
rowSamStatistics(x=x, y=y, resp.type = "Multiclass")

}
\author{
Alessandro Barberis
}
