% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-test-functions.R
\name{rowSpearmanCor}
\alias{rowSpearmanCor}
\title{Spearman's Correlation}
\usage{
rowSpearmanCor(
  x,
  y,
  alternative = c("two.sided", "greater", "less"),
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{x}{\code{matrix} or \code{data.frame}.}

\item{y}{numeric vector of data values. Must have the same length as \code{ncol(x)}.}

\item{alternative}{character string indicating the alternative hypothesis for
each row of \code{x}.
It must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"less"}.}

\item{conf.level}{numerical value or numeric vector of length \code{nrow(x)}.
The confidence levels of the intervals.
All values must be in the range of \eqn{[0:1]} or \code{NA}.}

\item{...}{further arguments to \code{\link[stats]{cor.test}}}
}
\value{
A list containing two elements:

\describe{
\item{statistic}{A numeric vector, the values of the test statistic}
\item{significance}{A numeric vector, the p-values of the selected test}
}
}
\description{
This function computes the \emph{Spearman's correlation} for each row
vector in \code{x}.
See the \strong{Details} section below for further information.
}
\details{
It is a wrapper to \code{\link[stats]{cor.test}}
function.
}
\examples{
#Seed
set.seed(1010)

#Data
x = matrix(rnorm(100 * 20), 100, 20)
y = rnorm(20)

#Compute
rowSpearmanCor(x = x, y = y)

}
\author{
Alessandro Barberis
}
