% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-test-functions.R
\name{rowTwoSampleT}
\alias{rowTwoSampleT}
\title{Two-sample t-Test}
\usage{
rowTwoSampleT(
  x,
  g,
  null = 0,
  alternative = c("two.sided", "greater", "less"),
  conf.level = 0.95,
  var = c("unequal", "equal"),
  paired = FALSE
)
}
\arguments{
\item{x}{\code{matrix} or \code{data.frame}.}

\item{g}{a vector or factor object giving the group for the corresponding
elements of \code{x}.}

\item{null}{numerical value or numeric vector of length \code{nrow(x)}.
The true values of the difference in means between the two groups of
observations for each row.}

\item{alternative}{character string or vector of length \code{nrow(x)}.
The alternative hypothesis for each row of \code{x}.
Values must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"less"}.}

\item{conf.level}{numerical value or numeric vector of length \code{nrow(x)}.
The confidence levels of the intervals.
All values must be in the range of \eqn{[0:1]} or \code{NA}.}

\item{var}{character string indicating whether to treat the two variances as
being unequal (default) or equal. If \code{var = "unequal"}, the Welch
approximation to the degrees of freedom is used.}

\item{paired}{logical, whether to compute a paired t-test.}
}
\value{
A list containing two elements:

\describe{
\item{statistic}{A numeric vector, the values of the test statistic}
\item{significance}{A numeric vector, the p-values of the selected test}
}
}
\description{
Computes the two-sample t-Test for each feature.

It internally calls functions from the\code{matrixTests}
package for the various tests.
}
\examples{
#Seed
set.seed(1010)

#Data
x = matrix(rnorm(100 * 20), 100, 20)
g = c(rep(0,10),rep(1,10))

#Compute two-sample t-test
rowTwoSampleT(x = x, g = g, var = "equal")

#Compute two-sample t-test with Welch approximation
rowTwoSampleT(x = x, g = g)

#Compute two-sample paired t-test
rowTwoSampleT(x = x, g = g, paired = TRUE)

}
\author{
Alessandro Barberis
}
