% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-test-functions.R
\name{rowUnequalVarT}
\alias{rowUnequalVarT}
\title{Welch's t-Test}
\usage{
rowUnequalVarT(
  x,
  g,
  null = 0,
  alternative = c("two.sided", "greater", "less"),
  conf.level = 0.95
)
}
\arguments{
\item{x}{\code{matrix} or \code{data.frame}.}

\item{g}{a vector or factor object giving the group for the corresponding
elements of \code{x}.}

\item{null}{numerical value or numeric vector of length \code{nrow(x)}.
The true values of the difference in means between the two groups of
observations for each row.}

\item{alternative}{character string or vector of length \code{nrow(x)}.
The alternative hypothesis for each row of \code{x}.
Values must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"less"}.}

\item{conf.level}{numerical value or numeric vector of length \code{nrow(x)}.
The confidence levels of the intervals.
All values must be in the range of \eqn{[0:1]} or \code{NA}.}
}
\value{
A list containing two elements:

\describe{
\item{statistic}{A numeric vector, the values of the test statistic}
\item{significance}{A numeric vector, the p-values of the selected test}
}
}
\description{
Computes the two-sample t-test with the Welch modification to the
degrees of freedom for each feature.

See the \strong{Details} section below for further information.
}
\details{
It is a wrapper to \code{\link[matrixTests]{row_t_welch}}
function.
}
\examples{
#Seed
set.seed(1010)

#Data
x = cbind(
 matrix(rnorm(n = 100 * 10, sd = 1), 100, 10),
 matrix(rnorm(n = 100 * 10, sd = 3), 100, 10)
)
g = c(rep(0,10),rep(1,10))

#Compute
rowUnequalVarT(x = x, g = g)

}
\author{
Alessandro Barberis
}
