% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-stats-functions.R
\name{rowVariability}
\alias{rowVariability}
\title{Variability}
\usage{
rowVariability(
  x,
  g = NULL,
  method = c("sd", "iqr", "mad", "rsd", "efficiency", "vmr")
)
}
\arguments{
\item{x}{\code{matrix} or \code{data.frame}, where rows are features and columns are observations.}

\item{g}{(optional) vector or factor object giving the group for the corresponding
elements of \code{x}.}

\item{method}{character string indicating the measure of variability.
Available options are:
\describe{
\item{\code{"sd"}}{the standard deviation}
\item{\code{"iqr"}}{the interquartile range}
\item{\code{"mad"}}{the median absolute deviation}
\item{\code{"rsd"}}{the relative standard deviation (i.e., coefficient of variation)}
\item{\code{"efficiency"}}{the coefficient of variation squared}
\item{\code{"vmr"}}{the variance-to-mean ratio}
}}
}
\value{
A \code{vector} of length \code{nrow(x)} containing the computed ratios.
If \code{g} is provided, a \code{matrix} with ratios for each class as column
vectors is returned.
}
\description{
This function computes measures of variability.
}
\details{
The corrected sample \emph{standard deviation} is the defined as:

\deqn{ SD = \sqrt{ \frac{1}{N-1} \sum_{i=1}{N}(x_{i} - \bar{x})^2 } }

where \eqn{x} is a vector of \eqn{N} elements, and \eqn{\bar{x}} is its mean value.

The \emph{interquartile range} is defined as the difference between the 75th and
25th percentiles of the data:

\deqn{ IQR = Q_{3} - Q_{1})}

The \emph{median absolute deviation} is defined as the median of the absolute
deviations from the mean:

\deqn{ MAD = \median(|x_{i} - \bar{x}|)}

The \emph{relative standard variation} is defined as the ratio of the standard deviation
to the mean:

\deqn{ RSD = \frac{standard deviation}{mean} = \frac{\sigma}{\mu}}

The \emph{efficiency} is the square of the coefficient of variation:

\deqn{ efficiency = (\frac{standard deviation}{mean})^2 = \frac{\sigma^2}{\mu^2}}

The \emph{variance-to-mean ratio} is computed as the ratio of the variance to the mean:

\deqn{ VMR = \frac{variance}{mean} = \frac{\sigma^2}{\mu}}

See also the following functions for further details:
\describe{
\item{\code{"sd"        }}{\code{\link{sd}}}
\item{\code{"iqr"       }}{\code{\link{iqr}}}
\item{\code{"mad"       }}{\code{\link{mad}}}
\item{\code{"rsd"       }}{\code{\link{rsd}}}
\item{\code{"efficiency"}}{\code{\link{efficiency}}}
\item{\code{"vmr"       }}{\code{\link{vmr}}}
}
}
\examples{
#Seed
set.seed(1010)

#Define row/col size
nr = 5
nc = 10

#Data
x = matrix(
 data = sample.int(n = 100, size = nr*nc, replace = TRUE),
 nrow = nr,
 ncol = nc,
 dimnames = list(
   paste0("f",seq(nr)),
   paste0("S",seq(nc))
 )
)

#Grouping variable
g = c(rep("a", nc/2), rep("b", nc/2))

#Variance-to-mean Ratio
rowVariability(x = x, method = 'vmr')

#Variance-to-mean Ratio by group
rowVariability(x = x, g = g, method = 'vmr')

}
\author{
Alessandro Barberis
}
