% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-stats-functions.R
\name{rsd}
\alias{rsd}
\title{Relative Standard Deviation (Coefficient of Variation)}
\usage{
rsd(x, g = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{numerical vector.}

\item{g}{(optional) vector or factor object giving the group for the corresponding
elements of \code{x}.}

\item{na.rm}{logical indicating whether missing values should be removed before
computation.}
}
\value{
A numerical value or a vector containing the computed measure per class.
}
\description{
This function computes the \emph{coefficient of variation}
of the values in \code{x}. It is also known as \emph{relative standard deviation}.

See the \strong{Details} section below for further information.
}
\details{
The \emph{coefficient of variation} - also known as \emph{Normalized}
\emph{Root-Mean-Square Deviation} (\emph{NRMSD}), \emph{Percent RMS}, and \emph{relative standard}
\emph{deviation} (\emph{RSD}) - is a measure of statistical dispersion.

It is defined as the ratio of the standard deviation to the mean:

\deqn{ CV = RSD = \frac{standard deviation}{mean} = \frac{\sigma}{\mu}}

It is be computed as:

\deqn{ RSD = \frac{ \sqrt{\sum_{j=1}^{n}} ({x}_{j} - \bar{x})^2 / n }{ \bar{x} } }

where \eqn{\bar{x}} is the sample mean.

If \code{x} can be partitioned into \eqn{c} subgroups (provided by \code{g}),
then the \eqn{RSD} is computed for each class.

The coefficient of variation ranges between \eqn{[-\infty, \infty]}.

The main advantage of the coefficient of variation is that it is unitless.
}
\examples{
#Seed
set.seed(1010)

#Define size
n = 10

#Data
x = sample.int(n = 100, size = n, replace = TRUE)

#Grouping variable
g = c(rep("a", n/2), rep("b", n/2))

#Coefficient of variation
rsd(x)

#Coefficient of variation by group
rsd(x = x, g = g)

}
\references{
Fong, Biuk-Aghai, Si, Lightweight Feature Selection Methods Based on
Standardized Measure of Dispersion for Mining Big Data,
2016 IEEE International Conference on Computer and Information Technology (CIT)
}
\author{
Alessandro Barberis
}
