% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-selection-functions.R
\name{selectByFdr}
\alias{selectByFdr}
\title{Select by False Discovery Rate}
\usage{
selectByFdr(x, alpha = 0.05, operator = c("<", "<="))
}
\arguments{
\item{x}{numerical vector indicating the p-values.}

\item{alpha}{numerical value, the upper-bound on the FDR.}

\item{operator}{character string indicating the relational operator to use.}
}
\value{
A logical vector, indicating the elements to keep.
}
\description{
Select elements based on the false discovery rate. The provided
significance (\code{x}) is corrected via the Benjamini-Hochberg procedure.
}
\examples{
#Seed
set.seed(1010)

#Data
x = seq(from = 0, to = 1, by = 0.1)
names(x) = paste0("f", seq(11))

#Select
selectByFdr(x = x, alpha = 0.3)

}
\author{
Alessandro Barberis
}
