% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-selection-functions.R
\name{selectByPercentile}
\alias{selectByPercentile}
\title{Select by Percentile}
\usage{
selectByPercentile(x, percentile = 0.25)
}
\arguments{
\item{x}{numerical vector.}

\item{percentile}{numerical value in the range \eqn{[0, 1]} indicating the
percentage of elements to keep.}
}
\value{
A logical vector, indicating the elements to keep.
}
\description{
Select elements based on the percentile of the highest values.
}
\examples{
#Seed
set.seed(1010)

#Data
x = seq(10)
names(x) = paste0("f", seq(10))

#Select
selectByPercentile(x = x, percentile = 0.25)

}
\author{
Alessandro Barberis
}
