% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-selection-functions.R
\name{selectByRanking}
\alias{selectByRanking}
\title{Select by Ranking}
\usage{
selectByRanking(
  x,
  k = length(x),
  method = c("auto", "shell", "quick", "radix"),
  decreasing = TRUE,
  use.abs = FALSE
)
}
\arguments{
\item{x}{numerical vector.}

\item{k}{integer value indicating the number of top elements to keep.}

\item{method}{character string indicating the method used for sorting (see
\code{\link[base]{sort}}).}

\item{decreasing}{logical indicating whether to sort in decreasing order.}

\item{use.abs}{logical indicating whether to use absolute values.}
}
\value{
A logical vector indicating the elements to keep.
}
\description{
Select elements based on their ranking.
}
\examples{
#Seed
set.seed(1010)

#Data
x = seq(10)
names(x) = paste0("f", seq(10))

#Select
selectByRanking(x = x, k = 5)

}
\author{
Alessandro Barberis
}
