% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-stats-functions.R
\name{snr}
\alias{snr}
\title{Signal-to-noise Ratio}
\usage{
snr(x, g = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{numerical vector.}

\item{g}{(optional) vector or factor object giving the group for the corresponding
elements of \code{x}.}

\item{na.rm}{logical indicating whether missing values should be removed before
computation.}
}
\value{
A numerical value or a vector containing the computed measure per class.
}
\description{
This function computes the \emph{signal-to-noise ratio}
of the values in \code{x}. It is the reciprocal of the coefficient of variation.

See the \strong{Details} section below for further information.
}
\details{
The \emph{signal-to-noise ratio} is here defined as the ratio of the mean to
the standard deviation:

\deqn{ SNR = \frac{mean}{standard deviation} = \frac{\mu}{\sigma} }

If \code{x} can be partitioned into \eqn{c} subgroups (provided by \code{g}),
then the \eqn{SNR} is computed for each class.
}
\examples{
#Seed
set.seed(1010)

#Define size
n = 10

#Data
x = sample.int(n = 100, size = n, replace = TRUE)

#Grouping variable
g = c(rep("a", n/2), rep("b", n/2))

#Signal-to-noise ratio
snr(x)

#Signal-to-noise ratio by group
snr(x = x, g = g)

}
\references{
https://en.wikipedia.org/wiki/Coefficient_of_variation
}
\author{
Alessandro Barberis
}
