% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-featscreen-class.R
\name{validateFeatscreen}
\alias{validateFeatscreen}
\title{Validator of \code{featscreen} Objects}
\usage{
validateFeatscreen(x)
}
\arguments{
\item{x}{a \code{\link{featscreen}} object.}
}
\value{
An object of class \code{featscreen}.
}
\description{
This is the validator of the \code{featscreen} objects.
}
\examples{
\dontrun{
#No error is raised
validateFeatscreen(newFeatscreen())

#An error is raised
validateFeatscreen(
newFeatscreen(
 method = 'cor.test',
 n = 3,
 keep = c(1, 3)
)
)
}
}
\author{
Alessandro Barberis
}
\keyword{internal}
