% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-stats-functions.R
\name{vmr}
\alias{vmr}
\title{Variance-to-mean Ratio}
\usage{
vmr(x, g = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{numerical vector.}

\item{g}{(optional) vector or factor object giving the group for the corresponding
elements of \code{x}.}

\item{na.rm}{logical indicating whether missing values should be removed before
computation.}
}
\value{
A numerical value or a vector containing the computed measure per class.
}
\description{
This function computes the \emph{variance-to-mean ratio}.

See the \strong{Details} section below for further information.
}
\details{
The \emph{variance-to-mean ratio} (also known as \emph{index of dispersion})
is a measure of dispersion. It is computed as:

\deqn{ VMR = \frac{variance}{mean} = \frac{\sigma^2}{\mu}}

If \code{x} can be partitioned into \eqn{c} subgroups (provided by \code{g}),
then the \eqn{VMR} is computed for each class.
}
\examples{
#Seed
set.seed(1010)

#Define size
n = 10

#Data
x = sample.int(n = 100, size = n, replace = TRUE)

#Grouping variable
g = c(rep("a", n/2), rep("b", n/2))

#Variance-to-mean ratio
vmr(x)

#Variance-to-mean ratio by group
vmr(x = x, g = g)

}
\references{
https://en.wikipedia.org/wiki/Index_of_dispersion
}
\author{
Alessandro Barberis
}
